ENG-01: extract PCA into _pca
Move the PCA estimator (~810 lines) into its own module; _pca_pls
re-imports PCA and drops the now PCA-only cross_val_score import.
